Remove BitcoinJsLibSwapProvider- Unify swap functionality under BitcoinSwapProvider
