+++ examples/swap/cli/bitcoin.js
@@ -8,7 +8,8 @@ chains.bitcoin = new Client()
 chains.bitcoin.addProvider(new providers.bitcoin.BitcoinBitcoreRpcProvider('http://localhost:18332', 'bitcoin', 'local321'))
 // chains.bitcoin.addProvider(new providers.bitcoin.BitcoinLedgerProvider({ network: networks.bitcoin_testnet, segwit: false }))
 // chains.bitcoin.addProvider(new providers.bitcoin.BitcoinSwapProvider({ network: networks.bitcoin_testnet }))
-chains.bitcoin.addProvider(new providers.bitcoin.BitcoinJsLibSwapProvider({ network: networks.bitcoin_testnet }))
+chains.bitcoin.addProvider(new providers.bitcoin.BitcoinNodeWalletProvider(networks.bitcoin_testnet, 'http://localhost:18332', 'bitcoin', 'local321'))
+chains.bitcoin.addProvider(new providers.bitcoin.BitcoinSwapProvider({ network: networks.bitcoin_testnet }))
 async function doSwap () {
   chains.bitcoin.swap.generateSecret('test').then(secret => {
     chains.bitcoin.wallet.getUnusedAddress().then(address => {
