--- packages/bitcoin-bitcoinjs-lib-swap-provider/lib/index.js
@@ -1,3 +1,0 @@
-import BitcoinBitcoinJsLibSwapProvider from './BitcoinBitcoinJsLibSwapProvider'
-
-export default BitcoinBitcoinJsLibSwapProvider
