--- packages/bitcoin-bitcoinjs-lib-swap-provider/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@liquality/bitcoin-bitcoinjs-lib-swap-provider",
-  "umdName": "BitcoinBitcoinJsLibSwapProvider",
-  "umdExport": "default",
-  "version": "0.2.11",
-  "description": "",
-  "module": "lib/index.js",
-  "main": "dist/index.cjs.js",
-  "files": [
-    "dist",
-    "lib"
-  ],
-  "scripts": {
-    "build:node": "webpack --config ../../webpack/webpack.node.config.js",
-    "build:browser": "webpack --config ../../webpack/webpack.browser.config.js",
-    "build": "webpack --config ../../webpack/webpack.config.js"
-  },
-  "author": "Liquality <info@liquality.io>",
-  "license": "MIT",
-  "dependencies": {
-    "@babel/runtime": "^7.4.3",
-    "@liquality/bitcoin-networks": "^0.2.11",
-    "@liquality/bitcoin-utils": "^0.2.11",
-    "@liquality/crypto": "^0.2.11",
-    "@liquality/provider": "^0.2.11",
-    "@liquality/utils": "^0.2.11",
-    "bignumber.js": "^9.0.0",
-    "bitcoinjs-lib": "^5.1.1"
-  },
-  "engines": {
-    "node": "~8.12.0"
-  },
-  "homepage": "https://github.com/liquality/chainabstractionlayer#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/liquality/chainabstractionlayer.git"
-  },
-  "bugs": {
-    "url": "https://github.com/liquality/chainabstractionlayer/issues"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "1855d9ec369c8c577f74eb04e310a7c00b9eaa6c"
-}
